refactor: split all helpers_test.go by meaning
